fixed msvc warnings

+++ src/frontend.cpp
@@ -307,8 +307,8 @@ static void tokenizeLaTeX(const char *str, const size_t len, vector<likely_ast> 
 {
     const char *begin = "\\begin{likely}";
     const char *end = "\\end{likely}";
-    const uint32_t beginLength = strlen(begin);
-    const uint32_t endLength = strlen(end);
+    const uint32_t beginLength = uint32_t(strlen(begin));
+    const uint32_t endLength = uint32_t(strlen(end));
     const uint32_t minLength = beginLength + endLength;
 
     size_t i = 0;
